+++ docs/logs/getting-started-aspnetcore/Program.cs
@@ -15,8 +15,6 @@ builder.Logging.ClearProviders();
 
 builder.Logging.AddOpenTelemetry(logging =>
 {
-    logging.IncludeScopes = true;
-
     var resourceBuilder = ResourceBuilder
         .CreateDefault()
         .AddService(builder.Environment.ApplicationName);
@@ -41,7 +39,7 @@ app.Logger.StartingApp();
 
 app.Run();
 
-public static partial class ApplicationLogs
+internal static partial class LoggerExtensions
 {
     [LoggerMessage(LogLevel.Information, "Starting the app...")]
     public static partial void StartingApp(this ILogger logger);