+++ src/OpenTelemetry.Instrumentation.AspNetCore/Implementation/HttpInListener.cs
@@ -38,7 +38,6 @@ internal class HttpInListener : ListenerHandler
     internal static readonly ActivitySource ActivitySource = new(ActivitySourceName, Version.ToString());
 
     private const string DiagnosticSourceName = "Microsoft.AspNetCore";
-    private const string UnknownHostName = "UNKNOWN-HOST";
 
     private static readonly Func<HttpRequest, string, IEnumerable<string>> HttpRequestHeaderValuesGetter = (request, name) =>
     {
@@ -53,11 +52,6 @@ internal class HttpInListener : ListenerHandler
 
     private static readonly PropertyFetcher<Exception> ExceptionPropertyFetcher = new("Exception");
 
-#if !NET6_0_OR_GREATER
-    private readonly PropertyFetcher<object> beforeActionActionDescriptorFetcher = new("actionDescriptor");
-    private readonly PropertyFetcher<object> beforeActionAttributeRouteInfoFetcher = new("AttributeRouteInfo");
-    private readonly PropertyFetcher<string> beforeActionTemplateFetcher = new("Template");
-#endif
     private readonly AspNetCoreTraceInstrumentationOptions options;
 
     public HttpInListener(AspNetCoreTraceInstrumentationOptions options)
@@ -106,7 +100,7 @@ internal class HttpInListener : ListenerHandler
         // By this time, samplers have already run and
         // activity.IsAllDataRequested populated accordingly.
 
-        HttpContext context = payload as HttpContext;
+        var context = payload as HttpContext;
         if (context == null)
         {
             AspNetCoreInstrumentationEventSource.Log.NullPayload(nameof(HttpInListener), nameof(this.OnStartActivity), activity.OperationName);
@@ -119,9 +113,10 @@ internal class HttpInListener : ListenerHandler
         if (textMapPropagator is not TraceContextPropagator)
         {
             var ctx = textMapPropagator.Extract(default, request, HttpRequestHeaderValuesGetter);
-
             if (ctx.ActivityContext.IsValid()
-                && ctx.ActivityContext != new ActivityContext(activity.TraceId, activity.ParentSpanId, activity.ActivityTraceFlags, activity.TraceStateString, true))
+                && !((ctx.ActivityContext.TraceId == activity.TraceId)
+                    && (ctx.ActivityContext.SpanId == activity.ParentSpanId)
+                    && (ctx.ActivityContext.TraceState == activity.TraceStateString)))
             {
                 // Create a new activity with its parent set from the extracted context.
                 // This makes the new activity as a "sibling" of the activity created by